improve development plugin, introduce core.caching, allow get-plugins to have loadOpts
