+++ .github/workflows/pr-core-tests.yml
@@ -10,7 +10,7 @@ jobs:
       TERM: xterm
       # DEBUG: "lando*"
       LANDO_CORE_RUNTIME: 4
-      LANDO_CORE_DEV__MODE: 1
+      LANDO_CORE_DEVELOPMENT: 1
       LANDO_CORE_TELEMETRY: 0
     strategy:
       fail-fast: false
