+++ plugins/development/package.json
@@ -1,4 +1,4 @@
 {
-  "name": "@lando/dev",
+  "name": "@lando/development",
   "description": "An optional core plugin to help with Lando development"
 }
