+++ plugins/development/tasks/registry.js
